requirements.txt
fastapi[standard]==0.112.1
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
sentry-sdk[fastapi]==2.13.0